% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_hollowpolygon_old.R
\name{geom_hollow_polygon2}
\alias{geom_hollow_polygon2}
\title{geom_hollowpolygon}
\usage{
geom_hollow_polygon2(mapping = NULL, data = NULL, stat = "hollow_contour",
  position = "identity", show.legend = NA, inherit.aes = TRUE, ...)
}
\description{
new geom to manage hollows in contour plot
}
\references{
See http://rpubs.com/kohske/352

             See http://docs.ggplot2.org/dev/vignettes/extending-ggplot2.html
}

