% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hollow_polygon.R
\docType{data}
\name{GeomHollowPolygon}
\alias{GeomHollowPolygon}
\title{geom_hollowpolygon}
\format{An object of class \code{GeomHollowPolygon} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomHollowPolygon
}
\description{
new geom to manage hollows in contour plot
}
\references{
See http://rpubs.com/kohske/352
See http://docs.ggplot2.org/dev/vignettes/extending-ggplot2.html
}
\keyword{datasets}
