% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_polygon2.R
\docType{data}
\name{GeomPolygon2}
\alias{GeomPolygon2}
\title{new geom
 See http://rpubs.com/kohske/3522
 See http://docs.ggplot2.org/dev/vignettes/extending-ggplot2.html}
\format{\preformatted{Classes 'GeomPolygon2', 'Geom', 'ggproto' <ggproto object: Class GeomPolygon2, Geom>
    default_aes: uneval
    drak_key: function
    draw: function
    draw_group: function
    draw_key: function
    draw_panel: function
    reparameterise: function
    required_aes: x y
    super:  <ggproto object: Class Geom> 
}}
\usage{
GeomPolygon2
}
\description{
new geom
 See http://rpubs.com/kohske/3522
 See http://docs.ggplot2.org/dev/vignettes/extending-ggplot2.html
}
\keyword{datasets}

