% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importADSOBIN.R
\name{importADSOBIN}
\alias{importADSOBIN}
\title{Import ADSO/BIN as raster}
\usage{
importADSOBIN(fname, k = 1, kz = 1, dx = 0, dy = 0,
  destaggering = FALSE, variable = NULL, slice = 1, deadline = 1,
  verbose = TRUE)
}
\arguments{
\item{fname}{(character) raster file to be imported.}

\item{k}{float. Factor to be applied to x and y coordinates (default = 1).}

\item{kz}{float. Factor to be applied to z values (default = 1).}

\item{dx}{float. Shift x coordinates by dx (default = 0).}

\item{dy}{float. Shift y coordinates by dy (default = 0).}

\item{destaggering}{logical. If `TRUE` destaggering is applied (default =
FALSE).}

\item{variable}{string, the name of the variable to be imported.}

\item{slice}{integer, horizontal slice for 3D variables (default = 1).}

\item{deadline}{intefer, temporal deadline (default = 1)}

\item{verbose}{logical. If `TRUE` print out basic statistics (default =
TRUE).}
}
\value{
A dataset with x, y and z columns is returned.
}
\description{
Import ADSO/BIN as raster
}
