\name{plotAvgRad}
\alias{plotAvgRad}
\title{Plot monthly average radiation}
\usage{
plotAvgRad(mydata, rad = "radg")
}
\arguments{
  \item{mydata}{A data frame containing fields date and
  radg}

  \item{radg}{Name of the column representing radiation}
}
\description{
Plot monthly average radiation
}

