\name{plotStabClass}
\alias{plotStabClass}
\title{Plot stability class}
\usage{
plotStabClass(mydata, type = "season")
}
\arguments{
  \item{mydata}{A data frame containing cs field}

  \item{type}{type determines how the data are split and
  then plotted. The default is "season". Any other string
  will give an hourly plot.}
}
\description{
Plot stability class
}

