% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityClass.R
\name{stabilityClass}
\alias{stabilityClass}
\title{Stability class.}
\usage{
stabilityClass(rad, tcc, ws, option = "iaea")
}
\arguments{
\item{rad}{The net radiation in W/m^2}

\item{tcc}{The total cloud cover in a range from 1  to 8}

\item{ws}{wind speed in m/s}

\item{option}{The method used to detrmine the stability class. It can be
\code{iaea} (default) to comply with ARIA Impact(tm), \code{pasquill} or \code{custom}.
Previous option \code{impact} is the same as \code{iaea} and it is now
deprecated.}
}
\value{
\code{stabilityClass} returns a numeric vector with Pasquill
stability classes coded as: A = 1, B = 2, ... , F = 6.
}
\description{
Computes stability class given net radiation, total cloud cover and wind
speed.
}
\details{
\code{stabilityClass()} computes stability class according to IAEA method based
on net radiation, total cloud cover tcc and wind speed.
Net radiation and wind are used by day; tcc and wind are used by night.

Three different alogorithms are implemented; see source code for details.
}
\examples{

# Compute stability class with custom algorithm
stMeteo$cst <- stabilityClass(rad = stMeteo$rad,
                              tcc = stMeteo$tcc,
                              ws = stMeteo$ws,
                              option = "custom")

}
\seealso{
\code{\link[=plotStabilityClass]{plotStabilityClass()}}
}
