% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_hollow_contour.R
\name{stat_hollow_contour}
\alias{stat_hollow_contour}
\title{Contour plot with hollow management}
\usage{
stat_hollow_contour(mapping = NULL, data = NULL, geom = "hollow_polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Contour plot with hollow management
}

