% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stMeteo}
\alias{stMeteo}
\title{Time series of meteorological variables}
\format{
A data frame with 8759 rows and 7 variables:
\describe{
\item{date}{Date and time of data collection (y-m-d h:m:s).}
\item{ws}{Wind speed (m/s).}
\item{wd}{Wind direction (deg).}
\item{temp}{Air temperature (Celsius).}
\item{radg}{Global radiation (W/m2).}
\item{tcc}{Total cloud cover (0-8).}
\item{pgt}{Pasquill-Gifford-Turner stability class (1-6).}
}
}
\usage{
stMeteo
}
\description{
A dataset containing time series of meteorological variables.
The variables are as follows:
}
\keyword{datasets}
