% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stTestmap}
\alias{stTestmap}
\title{Concentration values on a domain}
\format{
A data frame with 10201 rows and 3 variables.
\describe{
\item{x}{Easting coordinate (m).}
\item{y}{Northing coordinate (m).}
\item{z}{Concentration values to be plotted.}
}
}
\usage{
stTestmap
}
\description{
A dataset containing fake concentration values of an unspecified
airborne pollutant to test mapping functions.
The domain has size 101 x 101 and the coordinates are expressed in m (as if they were UTM coordinates.)
}
\keyword{datasets}
