% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorField.R
\name{vectorField}
\alias{vectorField}
\title{Vector field plot}
\usage{
vectorField(
  data,
  scalex = 1,
  scaley = 1,
  everyx = 1,
  everyy = 1,
  length = 0.1,
  size = 0.25
)
}
\arguments{
\item{data}{A dataframe containing data to be plotted in the form of: \emph{(x, y, u, v)}.}

\item{scalex}{length factor of vector \emph{x} component}

\item{scaley}{length factor of vector \emph{y} component}

\item{everyx}{keep one out of every \emph{everyx} values, along \emph{x} direction.}

\item{everyy}{keep one out of every \emph{everyy} values, along \emph{y} direction.}

\item{length}{arrow length.}

\item{size}{arrow size.}
}
\description{
\code{vectorField} plots a vector field of two \strong{velocities}.
}
\details{
This function plot a vector field given a data.frame with coordinates (x, y)
and corresponding velocities (u, v). Vectors are colored by magnitude (speed).

This function is heavily inspired by snippets of code in
\emph{R Graphics Cookbook} by Winston Chang (https://r-graphics.org/index.html).
}
\examples{
\dontrun{
metU <- importADSOBIN('/path/to/meteofile',
                      variable = 'U',
                      slice=2,
                      k = 1000,
                      verbose = TRUE)
metU <- as.data.frame(metU)
metU <- metU \%>\% 
        mutate(u = z, z = NULL)

metV <- importADSOBIN('/path/to/meteofile',
                      variable = 'V', 
                      slice=2, 
                      k = 1000, 
                      verbose = TRUE)
metV <- as.data.frame(metV)
metV <- metV \%>\%  
        mutate(v = z, z = NULL)

met <- merge(metU, metV, by = c("x", "y"))

vectorField(met, everyx = 2, everyy = 2, scalex = 10, scaley = 10) +
    coord_fixed(ratio = 1, xlim = c(0, 1000), ylim = c(0, 1000))

}
}
