% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourPlot.R
\name{contourPlot}
\alias{contourPlot}
\title{Contour plot of pollutant concentration}
\usage{
contourPlot(data, domain, background, underlayer, overlayer, legend = NULL,
  levels = NULL, transparency = 0.66)
}
\arguments{
\item{data}{A dataframe containing data to be plotted in the form of X, Y and Z (levels).}

\item{domain}{An array with min X, max X, min Y, max Y, number of ticks on X axis, number of ticks on Y axis (optional).}

\item{background}{String containing the path to the png file to be plotted as a basemap (optional).}

\item{overlayer}{Array of strings containing layers to be plotted on top of the contour plot (optional).}

\item{legend}{(string) Legend title (optional).}

\item{levels}{Array of levels for contour plot. If not set, automatic levels are plotted.}

\item{transparency}{Transparency level of the contour plot (between 0 and 1)}

\item{undelayer}{Array of strings containing layers to be plotted between basemap and contour plot (optional).}
}
\value{
A \code{ggplot2} plot
}
\description{
\code{contourPlot} plots a contour map of pollutants
}
\details{
This is a high level function to plot contour levels of a pollutants matrix.
}
\examples{
# Importa raster data
data <- importRaster(paste(dir, inputfile, sep=""), k = 1000, variable = "CONCAN") 

# Simple contour plot
contourPlot(data)

# Specifiy (sub)domain to be plotted; background image; legend title and pollutant levels.
contourPlot(data, domain(500000, 510000, 6000000, 6010000, 7, 7), background = "img/background.png", legend = "no2 [ug/m3]", levels = c(10, 20, 30, 40))

# Add underlayer (same for overlayer)
library(ggplot2)
library(maptools)
perimetro <- readShapeLines("path_to/perimetro.shp")
perimetro <- fortify(perimetro)
strada <- readShapeLines("path_to/strada.shp")
strada <- fortify(strada)
myUnderlayer <- vector(mode = "list", length = 2)
myUnderlayer[[1]] <- geom_polygon(data = perimetro, aes(long, lat, group = group), colour = "black", fill = NA, size = 0.1, alpha = 0.5)
myUnderlayer[[2]] <- geom_path(data = strada, aes(long, lat, group = group), colour = "grey", size = 0.1, alpha = 0.5)
contourPlot(data = test, background = "path_to/basemap.png", underlayer = myUnderlayer)

}

