% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_hollow_polygon.R
\docType{data}
\name{GeomHollowPolygon}
\alias{GeomHollowPolygon}
\title{geom_hollowpolygon}
\format{\preformatted{Classes 'GeomHollowPolygon', 'Geom', 'ggproto' <ggproto object: Class GeomHollowPolygon, Geom>
    aesthetics: function
    default_aes: uneval
    draw_group: function
    draw_key: function
    draw_layer: function
    draw_panel: function
    non_missing_aes: 
    parameters: function
    required_aes: x y
    setup_data: function
    use_defaults: function
    super:  <ggproto object: Class Geom> 
}}
\usage{
GeomHollowPolygon
}
\description{
new geom to manage hollows in contour plot
}
\references{
See http://rpubs.com/kohske/352

             See http://docs.ggplot2.org/dev/vignettes/extending-ggplot2.html
}
\keyword{datasets}

