% Generated by roxygen2 (4.0.1): do not edit by hand
\name{importRaster}
\alias{importRaster}
\title{Import generic raster file}
\usage{
importRaster(fname, k = 1, destaggering = FALSE)
}
\arguments{
\item{fname}{(character) raster file to be imported.}

\item{k}{integer. Factor to be applied to x and y coordinates (default = 1).}

\item{destaggering}{logical. If `TRUE` destaggering is applied (default = FALSE).}
}
\value{
A dataset with x, y and z columns is returned.
}
\description{
The function import the first layer of a generic raster file. Data are imported as an array of x, y, z columns.
X and y coordinats can be converted from km to m (default k=1000) and viceversa.
Destaggering is applied by default.
}
\details{
In output summery data are plotted.
}
\examples{
# Import binary grid file and convert coordinates from km to m, without destaggering:
mydata <- importSurferGrd("/path_to_file/filename.grd", k = 1000, destaggering = FALSE)
}

