% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgRad.R
\name{plotAvgRad}
\alias{plotAvgRad}
\title{Plot hourly average radiation}
\usage{
plotAvgRad(
  mydata,
  date = "date",
  rad = "radg",
  ylabel = NULL,
  title = "",
  locale = NULL
)
}
\arguments{
\item{mydata}{A data frame containing fields with solar radiation time
series.}

\item{date}{Name of the column representing date and time.}

\item{rad}{Name of the column representing radiation.}

\item{ylabel}{The label along the y axis.}

\item{title}{Optional plot title}

\item{local}{Locale to use for legend. Default is English, the only other
one currently supported is italian.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a histogram with hourly average of solar radiation, together with
hourly maxima for June and December.
}
\examples{
data(stMeteo)
plotAvgRad(stMeteo, date = "date", rad = "radg")

}
\seealso{
\code{\link[=plotStabilityClass]{plotStabilityClass()}}, \code{\link[=plotAvgTemp]{plotAvgTemp()}}
}
