% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotAvgRad}
\alias{plotAvgRad}
\title{Plot hourly average radiation}
\usage{
plotAvgRad(mydata, rad = "radg")
}
\arguments{
\item{mydata}{A data frame containing fields `date` and `rad`}

\item{rad}{Name of the column representing radiation}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a histogram with hourly average of solar radiation, together with June and December hourly averages (minimum and maximum).
}
\examples{
plotAvgRad(mydata)
plotAvgRad(mydata, rad="radg")
}

