% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgRad.R
\name{plotAvgRad}
\alias{plotAvgRad}
\title{Plot hourly average radiation}
\usage{
plotAvgRad(mydata, date = "date", rad = "radg")
}
\arguments{
\item{mydata}{A data frame containing fields \code{date} and \code{radg}}

\item{date}{Name of the column representing date and time}

\item{rad}{Name of the column representing radiation}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a histogram with hourly average of solar radiation, together with hourly maxima for June and December.
}
\examples{
\dontrun{
plotAvgRad(mydata)
plotAvgRad(mydata, rad="radg")
}
}
