% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStabilityClass.R
\name{plotStabilityClass}
\alias{plotStabilityClass}
\title{Plot stability class}
\usage{
plotStabilityClass(mydata, sc = "sc", type = "season", locale = NULL)
}
\arguments{
\item{mydata}{The input data frame, which should contain \code{date} and \code{stability class} fields.}

\item{sc}{Name of the column in the data frame that represents the stability class.}

\item{type}{Specify how the data are to be split and plotted. Accepted values are "season" (default) and "hour".}

\item{locale}{Choose the locale for day and month names. The current locale is used by default, but you can also
specify a different one from the supported locales listed in stringi::stri_locale_list().
All labels will be in English by default or in Italian if its locale is specified.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Histogram plot of stability classes by season or hour.
}
\details{
Numerical values of stability classes are mapped as: 1 = A, 2 = B, ..., 6 = F.
}
\examples{
data("stMeteo")

# Season plot of stability class pgt
plotStabilityClass(stMeteo, sc = "pgt", type = "season")

# Hourly plot of stability class pgt
plotStabilityClass(stMeteo, sc = "pgt", type = "hour")

# Override default locale
plotStabilityClass(stMeteo, sc = "pgt", type = "season", locale = "it_IT")

}
\seealso{
\code{\link[=stabilityClass]{stabilityClass()}}, \code{\link[=plotAvgRad]{plotAvgRad()}}, \code{\link[=plotAvgTemp]{plotAvgTemp()}}
}
