% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStabilityClass.R
\name{plotStabilityClass}
\alias{plotStabilityClass}
\title{Plot stability class}
\usage{
plotStabilityClass(mydata, sc = "sc", type = "season", locale = NULL)
}
\arguments{
\item{mydata}{A data frame containing \code{date} and \code{stability class}
fields.}

\item{sc}{The name of the stability class field.}

\item{type}{type determines how the data are split and then plotted.
Accepted values are "season" (default) and "hour".}

\item{locale}{Locale to use for day and month names. Default is current
locale. Supported locales are listed in stringi::stri_locale_list().
All other labels are in English by default or in Italian if its locale is
specified.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot histogram of stability class on season or hour base.
}
\details{
Numerical values of stability classes are mapped as: 1 = A, 2 = B, ..., 6 = F.
}
\examples{
data("stMeteo")

# Season plot of stability class pgt
plotStabilityClass(stMeteo, sc = "pgt", type = "season")

# Hourly plot of stability class pgt
plotStabilityClass(stMeteo, sc = "pgt", type = "hour")

# Override default locale
plotStabilityClass(stMeteo, sc = "pgt", type = "season", locale = "it_IT")

}
\seealso{
\code{\link[=stabilityClass]{stabilityClass()}}, \code{\link[=plotAvgRad]{plotAvgRad()}}, \code{\link[=plotAvgTemp]{plotAvgTemp()}}
}
