% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove data outliers}
\usage{
removeOutliers(x, k)
}
\arguments{
\item{x}{vector of data to be cleaned}

\item{k}{factor to applied to the interquartile range}
}
\value{
vector of clenaed data
}
\description{
Outliers are removed from a data set. The interquartile range is computed.
Data larger or smaller than k * IQR are removed.
}
\examples{
# Remove data outside 5 time the interquartile range
mydata$ws <- removeOutliers(mydata$ws, 5)


}
