% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollingMax.R
\name{rollingMax}
\alias{rollingMax}
\title{Compute rolling max}
\usage{
rollingMax(mydata, length = 24)
}
\arguments{
\item{mydata}{A vector of data}

\item{length}{The length of data subset where the maximum values has to be picked.}
}
\description{
The rolling maximum value along a series of data is computed.
}
\details{
It computes the maximum value centered alonga a subset of data.
}
\examples{
# Compute rolling max along 24 hours on hourly time serie
t$radg24h <- rollingMax(mydata=t$radg, length=24)

# Compute rolling max along 30 days on a hourly time serie
t$radg30d <- rollingMax(mydata=t$radg24h, length=24*30)

}

