\name{rollingMax}
\alias{rollingMax}
\title{Compute rolling max}
\usage{
rollingMax(mydata, length = 24)
}
\arguments{
  \item{mydata}{A vector of data}

  \item{length}{The length of data subset where the maximum
  values has to be picked.}
}
\description{
The maximum value along a series of data is computed.
}

